Ultimo aggiornamento (si spera).

+++ src/interfaces/Settings.java
@@ -139,8 +139,8 @@ public class Settings
 			rightDrop.draw( g );
 
 			int j;
-			int startX = (int) (leftLife.getMaxX() + (rightLife.getX() - leftLife.getMaxX())/2 - widthH*2);
-			int startY = (int) (leftLife.getY() + leftLife.getHeight()/2 - heightH/2);
+			float startX =  leftLife.getMaxX() + (rightLife.getX() - leftLife.getMaxX())/2 - widthH*2;
+			float startY = leftLife.getY() + leftLife.getHeight()/2 - heightH/2;
 			for(j = 0; j < vite/2; j++)
 				heart.draw( startX + widthH*j, startY, widthH, heightH );
 			if(vite%2 == 1)
